Mask is an array of integers now.
Also: cleaned up basins(...) arguments.

+++ python/basins_c.pxd
@@ -1,4 +1,4 @@
 # -*- mode: python -*-
 
 cdef extern from "../src/basins.hh":
-    bint basins(double *x, int Mx, double *y, int My, double *z, double *mask, bint output)
+    bint basins(double *x, int Mx, double *y, int My, double *z, int *mask, bint output)
